requirements.txt
waitress==2.1.2
mysqlclient==2.1.1
Flask-Cors==3.0.10
python-dotenv==1.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
SQLAlchemy-Utils==0.40.0